commands: bind diarize ms-durations and serve port to their real C++ width
Part of fuzz-values item 3's width audit (src/cli/mod.rs's Opt::int_bounds /
ValueType, committed separately):
- cmd_diarize.rs: --minimum-duration-ms/--merge-gap-ms bind to
  rac_diarization_options_t's int64_t fields, not the 32-bit default.
- cmd_serve.rs: --port binds to cmd_serve.cpp's uint16_t, via
  .int_bounds(0, 65535) so 70000/4294967296 are conversion errors like C++.

+++ src/commands/cmd_diarize.rs
@@ -227,12 +227,14 @@ pub fn register_diarize(app: &mut App) {
     );
     cmd.add_option(
         "--minimum-duration-ms,--min-duration",
-        ValueType::Int,
+        // rac_diarization_options_t::minimum_duration_ms is int64_t.
+        ValueType::Int64,
         "Drop segments shorter than this many ms (default 0)",
     );
     cmd.add_option(
         "--merge-gap-ms,--merge-gap",
-        ValueType::Int,
+        // rac_diarization_options_t::merge_gap_ms is int64_t.
+        ValueType::Int64,
         "Merge same-speaker segments closer than this many ms (default 0)",
     );
